move test to lz4_chunked_compression
